+++ src/ui/widgets/options-frame.lua
@@ -17,7 +17,7 @@ local Widgets = Addon.UserInterface.Widgets
   }
 ]]
 function Widgets:OptionsFrame(options)
-  local BUTTONS_PER_ROW = 3
+  local BUTTONS_PER_ROW = 4
   local SPACING = Widgets:Padding()
 
   -- Defaults.
@@ -106,8 +106,8 @@ function Widgets:OptionButton(options)
   -- Label text.
   frame.label = frame:CreateFontString("$parent_Label", "ARTWORK", "GameFontNormal")
   frame.label:SetText(Colors.White(options.labelText))
-  frame.label:SetPoint("LEFT", frame, self:Padding(0.5), 0)
-  frame.label:SetPoint("RIGHT", frame, -self:Padding(0.5), 0)
+  frame.label:SetPoint("LEFT", frame, self:Padding(), 0)
+  frame.label:SetPoint("RIGHT", frame, -self:Padding(), 0)
   frame.label:SetWordWrap(false)
 
   function frame:UpdateTooltip()