+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 ## [Unreleased]
 
+### Added
+
+- Option: `Include Unsuitable Equipment`
+
 ### Changed
 
 - SavedVariables now populate/depopulate default values on login/logout
+- Made some minor UI modifications
+
+### Fixed
+
+- Fixed bug with initial slider values in the ItemsFrame
 
 ## [1.0.2] - 2022-09-26